feat(): add route handler for creating martial art
